update  refactor cod, remove combineReducers style
