+++ smurfs/src/components/Home.js
@@ -1,0 +1,13 @@
+import React from "react";
+import SmurfList from "./SmurfList";
+
+const Home = () => {
+  return (
+    <>
+      <h1>Welcome to the Smurf Management System, powered by React/ Redux!</h1>
+      <SmurfList />
+    </>
+  );
+};
+
+export default Home;
