--- smurfs/src/components/Login.js
@@ -1,2 +1,0 @@
-import React, { useState } from "react";
-import { axiosWithAuth } from "../";
